+++ packages/component-library-nuxt-module/CHANGELOG.md
@@ -1,3 +1,11 @@
+# [@ucla-library/component-library-nuxt-module-v1.4.19](https://github.com/UCLALibrary/ucla-library-website-components/compare/@ucla-library/component-library-nuxt-module-v1.4.18...@ucla-library/component-library-nuxt-module-v1.4.19) (2026-09-15)
+
+# [@ucla-library/component-library-nuxt-module-v1.4.18](https://github.com/UCLALibrary/ucla-library-website-components/compare/@ucla-library/component-library-nuxt-module-v1.4.17...@ucla-library/component-library-nuxt-module-v1.4.18) (2026-09-09)
+
+# [@ucla-library/component-library-nuxt-module-v1.4.17](https://github.com/UCLALibrary/ucla-library-website-components/compare/@ucla-library/component-library-nuxt-module-v1.4.16...@ucla-library/component-library-nuxt-module-v1.4.17) (2026-09-09)
+
+# [@ucla-library/component-library-nuxt-module-v1.4.16](https://github.com/UCLALibrary/ucla-library-website-components/compare/@ucla-library/component-library-nuxt-module-v1.4.15...@ucla-library/component-library-nuxt-module-v1.4.16) (2026-09-04)
+
 # [@ucla-library/component-library-nuxt-module-v1.4.15](https://github.com/UCLALibrary/ucla-library-website-components/compare/@ucla-library/component-library-nuxt-module-v1.4.14...@ucla-library/component-library-nuxt-module-v1.4.15) (2026-09-04)
 
 # [@ucla-library/component-library-nuxt-module-v1.4.14](https://github.com/UCLALibrary/ucla-library-website-components/compare/@ucla-library/component-library-nuxt-module-v1.4.13...@ucla-library/component-library-nuxt-module-v1.4.14) (2026-09-01)